Divide index.jsx into components
